Add files via upload
When working with scene strips in the VSE, Blender by default doesn't give any feedback as to which frame number of a scene strip is under the playhead. Instead, it mostly displays information about the strip in the context of the sequence, which isn't particularly useful.

Knowing the "real" frame number of a scene strip is vital should you want make changes to the internal scene at the specific frame beneath the playhead, or between a specific range of frames i.e. the frames between the in and out point of the strip.

For that reason I have coded a simple panel in Python that serves this precise function (see below). I'm not a full fledged developer really, so I'm hoping one of the Blender developers or someone in the community will take this code and add it to the next version of Blender or fork and turn it into a proper addon. If you do please make a fork here and make it available as GPL.